refactor(evaluatorq-py): move fields helper to common/ (RES-897)
